refactor(account-setup): extract SignatureContent to separate file
